Move Event and Reporter out to execute package
