+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/execute/executor/SyncJobExecutor.java
@@ -20,6 +20,7 @@ package info.avalon566.shardingscaling.core.execute.executor;
 import info.avalon566.shardingscaling.core.config.SyncConfiguration;
 import info.avalon566.shardingscaling.core.controller.ReportCallback;
 import info.avalon566.shardingscaling.core.controller.SyncTaskProgress;
+import info.avalon566.shardingscaling.core.execute.Reporter;
 
 import java.util.List;
